move WorkingArea to its own file
